deployments/master/docker-img/requirements.txt
kafka-python
numpy
